+++ gtkwave3/src/search.c
@@ -181,7 +181,7 @@ int i;
 
 for(i=0;i<5;i++) GLOBALS->regex_mutex_search_c_1[i]=0;
 
-GLOBALS->regex_which_search_c_1=(int)((long)which);
+GLOBALS->regex_which_search_c_1=(int)((intptr_t)which);
 GLOBALS->regex_mutex_search_c_1[GLOBALS->regex_which_search_c_1] = 1; /* mark our choice */
 
 
@@ -1095,7 +1095,7 @@ void searchbox(char *title, GtkSignalFunc func)
         group = gtk_radio_menu_item_group (GTK_RADIO_MENU_ITEM (GLOBALS->menuitem_search[i]));
         gtk_menu_append (GTK_MENU (menu), GLOBALS->menuitem_search[i]);
         gtk_widget_show (GLOBALS->menuitem_search[i]);
-        gtkwave_signal_connect(GTK_OBJECT (GLOBALS->menuitem_search[i]), "activate", GTK_SIGNAL_FUNC(regex_clicked), (void *)((long)i));
+        gtkwave_signal_connect(GTK_OBJECT (GLOBALS->menuitem_search[i]), "activate", GTK_SIGNAL_FUNC(regex_clicked), (void *)((intptr_t)i));
         GLOBALS->regex_mutex_search_c_1[i]=0;
         }
